queue-proxy: Add customizable target IP
When the network setup is more complicated than a single NETNS talking to
localhost (eg. when running the user-container as a sandboxed container
or unikernel), we need to be able to customize the queue-proxy target IP.
For our use-case, we do that through the container runtime, appending an
environment variable to the queue-proxy container.

Signed-off-by: Ioannis Plakas <iplakas@nubificus.co.uk>
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>